# backend/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Backend API
  description: API documentation for the backend service
  version: "1.0.0"
servers:
  - url: http://localhost:8888
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    ResponseObject:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          nullable: true
    ResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...
    ImportReceiptRequest:
      type: object
      required: [supplierId, items]
      properties:
        receiptCode:
          type: string
        supplierId:
          type: integer
        note:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ImportItemRequest"
    ImportItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
          format: decimal
        unitPrice:
          type: number
          format: decimal
        warrantyMonths:
          type: integer
        serialNumbers:
          type: array
          items:
            type: string
        locationId:
          type: integer
    ImportReceiptResponse:
      type: object
      properties:
        id:
          type: integer
        receiptCode:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        totalAmount:
          type: number
          format: decimal
        status:
          type: string
          enum: [PENDING, PENDING_APPROVAL, COMPLETED, CANCELLED]
        note:
          type: string
          nullable: true
        createdBy:
          type: integer
        createdByName:
          type: string
        approvedBy:
          type: integer
          nullable: true
        approvedByName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ImportItemResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ImportItemResponse:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: number
          format: decimal
        unitPrice:
          type: number
          format: decimal
        warrantyMonths:
          type: integer
        createdUnits:
          type: integer

    # =========================== EXPORT RECEIPT ===========================
    ExportReceiptRequest:
      type: object
      required: [reason, items]
      properties:
        reason:
          type: string
          enum: [SALE, INTERNAL, RETURN_SUPPLIER, DISPOSE]
        customerId:
          type: integer
        note:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExportItemRequest"
    ExportItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
          format: decimal
        unitPrice:
          type: number
          format: decimal
    ExportReceiptResponse:
      type: object
      properties:
        id:
          type: integer
        receiptCode:
          type: string
        reason:
          type: string
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        totalAmount:
          type: number
          format: decimal
        status:
          type: string
          enum: [PENDING_APPROVAL, COMPLETED, CANCELLED]
        note:
          type: string
          nullable: true
        createdBy:
          type: integer
        createdByName:
          type: string
        approvedBy:
          type: integer
          nullable: true
        approvedByName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExportItemResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExportItemResponse:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: number
          format: decimal
        unitPrice:
          type: number
          format: decimal

    # =========================== PRODUCT UNIT ===========================
    ProductUnitResponse:
      type: object
      properties:
        id:
          type: integer
        serialNumber:
          type: string
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        trackingType:
          type: string
          enum: [SERIALIZED, BULK]
        initialQuantity:
          type: number
          format: decimal
        remainingQuantity:
          type: number
          format: decimal
        importReceiptItemId:
          type: integer
        locationId:
          type: integer
          nullable: true
        locationCode:
          type: string
          nullable: true
        status:
          type: string
        importedAt:
          type: string
          format: date-time
        warrantyMonths:
          type: integer
        warrantyStartDate:
          type: string
          format: date-time
          nullable: true
        warrantyExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================== STOCK CHECK ===========================
...
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/import-receipt:
    post:
      tags: [Import Receipt]
      summary: Create a new import receipt
      operationId: createImportReceipt
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportReceiptRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ImportReceiptResponse"

    get:
      tags: [Import Receipt]
      summary: List import receipts (paginated)
      operationId: listImportReceipts
      security:
        - BearerAuth: []
      parameters:
        - name: receiptCode
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/import-receipt/{id}:
    get:
      tags: [Import Receipt]
      summary: Get import receipt by ID
      operationId: getImportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ImportReceiptResponse"
  /api/v1/import-receipt/{id}/approve:
    put:
      tags: [Import Receipt]
      summary: Approve import receipt
      operationId: approveImportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
  /api/v1/import-receipt/{id}/cancel:
    put:
      tags: [Import Receipt]
      summary: Cancel import receipt
      operationId: cancelImportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
  /api/v1/import-receipt/{id}/units:
    get:
      tags: [Import Receipt]
      summary: Get product units in an import receipt
      operationId: getImportReceiptUnits
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductUnitResponse"

  # =========================== EXPORT RECEIPT ===========================
  /api/v1/export-receipt:
    post:
      tags: [Export Receipt]
      summary: Create a new export receipt
      operationId: createExportReceipt
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportReceiptRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ExportReceiptResponse"

    get:
      tags: [Export Receipt]
      summary: List export receipts (paginated)
      operationId: listExportReceipts
      security:
        - BearerAuth: []
      parameters:
        - name: receiptCode
          in: query
          schema:
            type: string
        - name: reason
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/export-receipt/{id}:
    get:
      tags: [Export Receipt]
      summary: Get export receipt by ID
      operationId: getExportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ExportReceiptResponse"
  /api/v1/export-receipt/{id}/approve:
    put:
      tags: [Export Receipt]
      summary: Approve export receipt
      operationId: approveExportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
  /api/v1/export-receipt/{id}/cancel:
    put:
      tags: [Export Receipt]
      summary: Cancel export receipt
      operationId: cancelExportReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"

  # =========================== PRODUCT UNIT ===========================
  /api/v1/product-unit:
    get:
      tags: [Product Unit]
      summary: List product units with filters (paginated)
      operationId: listProductUnits
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: productName
          in: query
          schema:
            type: string
        - name: serialNumber
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/product-unit/{id}:
    get:
      tags: [Product Unit]
      summary: Get product unit by ID
      operationId: getProductUnit
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ProductUnitResponse"
  /api/v1/product-unit/status/{status}:
    get:
      tags: [Product Unit]
      summary: List product units by status
      operationId: getProductUnitsByStatus
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/product-unit/product/{productId}:
    get:
      tags: [Product Unit]
      summary: List product units by product
      operationId: getProductUnitsByProduct
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"

  # =========================== STOCK CHECK ===========================
...